Main.tsx: moved content of the Articles page to Main page. Articles.tsx: removed Articles page component. Main.test.tsx: Articles.test.tsx renamed to Main.test.tsx and moved to src/pages/Main. Main.test.tsx: adjusted to work with Main.tsx component. Main.module.scss: renamed from Articles.module.scss to Main.module.scss. constants.ts: removed Articles page data from NAVIGATION constant. constants.ts: removed DUMMY_CHATS and STYLES unused constants. App.tsx: removed Articles page from routing
