+++ src/pages/Main/Main.module.scss
@@ -1,0 +1,23 @@
+.wrapper {
+  max-width: 1144px;
+  margin: 0 auto;
+}
+
+.articles-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 25px;
+  padding: 30px;
+  list-style-type: none;
+  justify-content: space-evenly;
+
+  &__item {
+    flex: 0 1 300px;
+    background-color: #E6F5FF;
+    padding: 15px;
+
+    &__img {
+      max-width: 300px;
+    }
+  }
+}
