+++ src/pages/Main/Main.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { Articles } from 'src/pages/Articles/Articles';
+import { Main } from 'src/pages/Main/Main';
 import { configureStore } from '@reduxjs/toolkit';
 import { articlesReducer } from 'store/articles/slice';
 
@@ -14,7 +14,7 @@ describe('Articles', () => {
   it('should render', () => {
     render(
       <Provider store={store}>
-        <Articles />
+        <Main />
       </Provider>
     );
   });
@@ -22,7 +22,7 @@ describe('Articles', () => {
   it('should render with snapshot', () => {
     const { asFragment } = render(
       <Provider store={store}>
-        <Articles />
+        <Main />
       </Provider>
     );
